Move graphql queries, use fragments and protect against unpubbed transits
